requirements.txt
discord.py>=2.4
easyocr
Pillow
deep-translator
numpy
python-dotenv
aiohttp
